+++ packages/query/src/utils/asyncInterval.ts
@@ -5,7 +5,7 @@
 
 const asyncIntervals: {run: boolean; id: number | NodeJS.Timeout}[] = [];
 
-const runAsyncInterval = async (cb, interval, intervalIndex) => {
+const runAsyncInterval = async (cb: () => any, interval: number, intervalIndex: number) => {
   await cb();
   if (asyncIntervals[intervalIndex].run) {
     // eslint-disable-next-line @typescript-eslint/no-misused-promises
@@ -13,20 +13,13 @@ const runAsyncInterval = async (cb, interval, intervalIndex) => {
   }
 };
 
-export const setAsyncInterval = (cb, interval) => {
+export const setAsyncInterval = (cb: () => any, interval: number) => {
   if (cb && typeof cb === 'function') {
     const intervalIndex = asyncIntervals.length;
     asyncIntervals.push({run: true, id: 0});
-    runAsyncInterval(cb, interval, intervalIndex);
+    void runAsyncInterval(cb, interval, intervalIndex);
     return intervalIndex;
   } else {
     throw new Error('Callback must be a function');
-  }
-};
-
-const clearAsyncInterval = (intervalIndex) => {
-  if (asyncIntervals[intervalIndex].run) {
-    clearTimeout(asyncIntervals[intervalIndex].id);
-    asyncIntervals[intervalIndex].run = false;
   }
 };